backend/requirements.txt
fastapi[standard]
uvicorn
ibm-watsonx-ai
python-dotenv
langchain
chromadb
pydantic
sqlalchemy
langchain-ibm
langchain_community
pypdf